pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "anyio>=4.12.1",
    "backoff>=2.2.1",
    "black>=26.1.0",
    "chardet>=5.0.0",
    "click>=8.3.1",
    "fastapi>=0.128.0",
    "gunicorn>=25.0.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "jsonschema>=4.26.0",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.8",
    "langchain-text-splitters>=1.1.0",
    "langfuse>=3.12.1",
    "litellm>=1.81.7",
    "mypy>=1.19.1",
    "numpy>=2.4.2",
    "openai>=2.16.0",
    "pandas>=3.0.0",
    "prometheus-client>=0.24.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=3.17.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.16.2",
    "rich>=14.3.2",
    "ruff>=0.15.0",
    "sentence-transformers>=5.2.2",
    "sentry-sdk>=2.51.0",
    "sqlalchemy>=2.0.46",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "tokenizers>=0.22.2",
    "tqdm>=4.67.3",
    "trafilatura>=1.6.0",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-benchmark>=5.2.3",
    "pytest-cov>=7.0.0",
    "pytest-httpx>=0.36.0",
    "ruff>=0.15.0",
]
